finished homework 8. feat: add startTime/duration validation for tasks
