+++ src/exceptions/TaskIntersectException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class TaskIntersectException extends RuntimeException {
+    public TaskIntersectException(String message) {
+        super(message);
+    }
+}
